feat: changing type quantity to decimal

+++ Models/ItemModel.cs
@@ -3,7 +3,7 @@ namespace Crud.Models;
 public class ItemModel
 {
     public int Id { get; set; }
-    public required int Quantity { get; set; }
+    public required decimal Quantity { get; set; }
     public required string UnitMeasure { get; set; }
     public int ProductId { get; set; }
     public required ProductModel Product { get; set; }
